viewinventory is working
