+++ ART_Gallery/Views/Employee/ShowInventory.cshtml
@@ -1,0 +1,27 @@
+﻿@model IEnumerable<ART_Gallery.ViewModels.MainViewModel>
+
+<h2>Inventory</h2>
+<table>
+    @foreach (var item in Model)
+    {
+        <tr>
+            <td>
+                @Html.DisplayFor(modelItem => item.Name)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.Inventory)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.RetailPrice)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.PurchasePrice)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.Image)
+            </td>
+        </tr>
+    }
+</table>
+
+
